Added more classes to BaseClass test
