+++ apps/site/models/abstract/mixins.py
@@ -273,3 +273,65 @@ class BaseMediaMixin(models.Model):
     @classmethod
     def make_directory(cls, path):
         upload_helpers.make_directory(path)
+
+
+class ObjectPermissionsMixin(models.Model):
+    from django.contrib.contenttypes import fields
+    """
+    Abstract base class for media groups (Project and View objects).
+    """
+    access_authority = models.ForeignKey('ObjectAuthority',
+                                         db_column='view_authority',
+                                         verbose_name='Sharing')
+    access_key = models.CharField(max_length=16, null=True, blank=True)
+    users = fields.GenericRelation('UserAuthorityObject')
+
+    class Meta:
+        abstract = True
+        app_label = 'site'
+
+    def __has_user_permissions(self, user, authority_id):
+        # anonymous or null users don't have user-level permissions:
+        if user is None or not user.is_authenticated():
+            return False
+
+        # object owners have blanket view/edit/manage user-level permissions:
+        if self.owner == user:
+            return True
+
+        # users with privileges which are greater than or equal to
+        # the authority_id have user-level permisisons:
+        return len(self.users
+                   .filter(user=user)
+                   .filter(authority__id__gte=authority_id)
+                   ) > 0
+
+    def can_view(self, user=None, access_key=None):
+        from localground.apps.site.models.permissions import ObjectAuthority, \
+            UserAuthority
+        # projects and views marked as public are viewable:
+        if self.access_authority.id == ObjectAuthority.PUBLIC:
+            return True
+
+        # projects and views marked as "PUBLIC_WITH_LINK" that provide
+        # the correct access_key are viewable:
+        elif self.access_authority.id == ObjectAuthority.PUBLIC_WITH_LINK \
+                and self.access_key == access_key:
+            return True
+
+        # projects which are accessible by the user are viewable:
+        else:
+            return self.__has_user_permissions(user, UserAuthority.CAN_VIEW)
+
+    def can_edit(self, user):
+        from localground.apps.site.models.permissions import UserAuthority
+        return self.__has_user_permissions(user, UserAuthority.CAN_EDIT)
+
+    def can_manage(self, user):
+        from localground.apps.site.models.permissions import UserAuthority
+        return self.__has_user_permissions(user, UserAuthority.CAN_MANAGE)
+
+    def share_url(self):
+        return '/profile/{0}/{1}/share/'.format(
+            self.model_name_plural,
+            self.id)
